% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutadapt_helper.R
\name{allOrients}
\alias{allOrients}
\title{all possible orientations of a DNA string
Code was pulled from dada2 document https://benjjneb.github.io/dada2/ITS_workflow.html
Benjamin Callahan gets all the credit}
\usage{
allOrients(primer)
}
\arguments{
\item{primer}{DNA sequence}
}
\value{
all possible orientations of a DNA sequence (forward, reverse, complement, reverse complement)
}
\description{
all possible orientations of a DNA string
Code was pulled from dada2 document https://benjjneb.github.io/dada2/ITS_workflow.html
Benjamin Callahan gets all the credit
}
