% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{assess_example}
\alias{assess_example}
\title{Larger example dataset for further runs}
\format{
\subsection{\code{assess_example}}{

A list with of three tibbles:
\describe{
\item{metadata}{Example metadata tibble}
\item{asvtable}{Example ASV count table}
\item{taxa}{Example ASV taxonomy table}

...
}
}
}
\source{
made up data following typical dada2 ASV count and taxonomy table style
}
\usage{
assess_example
}
\description{
Sufficient for assess_run()
}
\keyword{datasets}
