% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_run_copy.R
\name{assess_run_copy}
\alias{assess_run_copy}
\title{Assess sequencing results}
\usage{
assess_run_copy(
  metadata = NULL,
  asvtable = NULL,
  wells = "Well",
  plate = NULL,
  category = NULL,
  minReadCount = 0,
  pcoa = FALSE,
  corrplot = FALSE,
  ...
)
}
\arguments{
\item{metadata}{metadata object}

\item{asvtable}{asvtable object}

\item{wells}{wells for each sample. Expecting A01, A02, etc.}

\item{plate}{plate samples were on. Can be left as NULL if there's only one plate.}

\item{category}{Varible to label wells by. Makes the most sense if positive and negative controls are labeled.}

\item{minReadCount}{Minimum number of total reads for sample to be kept. Recommend to keep it at 0, but can change that if you want.}

\item{pcoa}{Include PCoA data (can significantly increase run time, so set to FALSE by default)}

\item{corrplot}{Generate correlation plots. Set to FALSE by default as for now not sure how helpful it actually is}

\item{...}{Allows passing of arguments to nested functions. not being used at the moment.}
}
\value{
figures visualizing results
}
\description{
Assess sequencing results
}
\examples{
out <- micro4R::assess_example
met <- out$metadata
asv <- out$asvtable
assess_run_copy(metadata = met, asvtable = asv, wells = "well", plate = "Plate", category = "SampleType")
}
