% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAlpha.R
\name{calcAlpha}
\alias{calcAlpha}
\title{Title}
\usage{
calcAlpha(
  asvtable = NULL,
  metadata = NULL,
  numRare = 400,
  method = "standard",
  category = NULL,
  minReads = 1000
)
}
\arguments{
\item{asvtable}{ASV table}

\item{metadata}{metadata}

\item{numRare}{optional number of rarefactions to perform on data}

\item{method}{doesn't do anything yet; may remove}

\item{category}{category to use for grouping data}

\item{minReads}{minimum number of reads to keep a sample}
}
\value{
a list of
}
\description{
Title
}
\examples{
asvtable <- unsampled_example()$asvtable
metadata <- unsampled_example()$metadata
calcAlpha(asvtable = asvtable, metadata = metadata, category = "SampleType")
}
