% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBeta.R
\name{calcBeta}
\alias{calcBeta}
\title{Calculate beta diversity dissimilarities with or without rarefaction}
\usage{
calcBeta(
  asvtable = NULL,
  metadata = NULL,
  numRare = 400,
  method = "bray",
  category = NULL,
  minReads = 1000,
  ...
)
}
\arguments{
\item{asvtable}{Input ASV table. Can be filtered with filtering()}

\item{metadata}{Metadata input object}

\item{numRare}{Number of rarefactions to perform. Set to 0 if you do not want to do any rarefication.}

\item{method}{Dissimilarity index to use. See ?vegan::vegdist for more information.}

\item{category}{Name in metadata column that you'd like to use}

\item{minReads}{Minimum number of reads to keep a sample}

\item{...}{allow passed arguments to nested functions}
}
\value{
A dissimilarity matrix
}
\description{
Calculate beta diversity dissimilarities with or without rarefaction
}
\examples{
asvtable <- unsampled_example()$asvtable
metadata <- unsampled_example()$metadata
calcBeta(asvtable = asvtable, metadata = metadata, category = "host_sex", method = "bray")
}
