% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro4R-utils.R
\name{checkAll}
\alias{checkAll}
\title{checker utils in this section
Check your ASV count and taxonomy tables, and optionall metadata, for potential issues.}
\usage{
checkAll(asvtable = NULL, taxa = NULL, metadata = NULL, ids = "SampleID")
}
\arguments{
\item{asvtable}{The ASV table}

\item{taxa}{Its associated taxonomy table}

\item{metadata}{The associated metadata table}

\item{ids}{SampleIDs}
}
\value{
Notifies user of issues with asvtable, taxonomy table, or metadata that could cause downstream problems.
}
\description{
checker utils in this section
Check your ASV count and taxonomy tables, and optionall metadata, for potential issues.
}
\examples{
out <- dada2_wrapper(example = TRUE)
checkAll(asvtable = out$asvtable, taxa = out$taxa, metadata = out$metadata)

}
