% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro4R-utils.R
\name{checkMeta}
\alias{checkMeta}
\title{Check metadata file to identify potential downstream issues.}
\usage{
checkMeta(df, ids = "SampleID")
}
\arguments{
\item{df}{A data frame or tibble containing your sample metadata.}

\item{ids}{The column name in your data frame that identifies the sample IDs.}
}
\value{
Returns warnings or errors with your metadata object that may cause downstream problems.
}
\description{
Check metadata file to identify potential downstream issues.
}
\examples{
metadata <- data.frame(
  SampleIDs = c("Sample1", "Sample2", "Sample3"),
  Age       = c(34, 58, 21),
  Health    = c("Healthy", "Sick", NA)
)
checkMeta(metadata, "SampleIDs")
}
