% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro4R-utils.R
\name{checkSampleID}
\alias{checkSampleID}
\title{Simple function to check for a column called SampleID}
\usage{
checkSampleID(df)
}
\arguments{
\item{df}{Your ASV table or metadata object}
}
\value{
Your ASV table or metadata object with your sample IDs column named 'SampleID'
}
\description{
Simple function to check for a column called SampleID
}
\examples{
checkSampleID(example_metadata())
}
