% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro4R-utils.R
\name{contaminate}
\alias{contaminate}
\title{"Contaminate" the example ASV table so decontam can run}
\usage{
contaminate()
}
\value{
An artificially "contaminated" ASV table
}
\description{
"Contaminate" the example ASV table so decontam can run
}
\examples{
contaminated_asvtable <- contaminate()
}
