% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutadapt_helper.R
\name{cutadapt_helper}
\alias{cutadapt_helper}
\title{Generate a shell script to run cutadapt to remove primer sequences}
\usage{
cutadapt_helper(
  example = FALSE,
  where = NULL,
  FWD = "CCTACGGGNGGCWGCAG",
  REV = "GACTACHVGGGTATCTAATCC",
  patternF = "R1_001.fastq.gz",
  patternR = "R2_001.fastq.gz",
  chatty = TRUE
)
}
\arguments{
\item{example}{set to TRUE to run the example}

\item{where}{where to look for your fastq files}

\item{FWD}{forward primer sequence}

\item{REV}{reverse primer sequence}

\item{patternF}{pattern to identify the forward reads}

\item{patternR}{pattern to identify the reverse reads}

\item{chatty}{how verbose to be}
}
\value{
a shell script for you to run to remove adapter/primer sequences
}
\description{
Generate a shell script to run cutadapt to remove primer sequences
}
