% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_asvtable.R
\name{dada2_asvtable}
\alias{dada2_asvtable}
\title{dada2 workflow to generate ASV table}
\usage{
dada2_asvtable(
  where = NULL,
  example = FALSE,
  patternF = "_R1_001.fastq.gz",
  patternR = "_R2_001.fastq.gz",
  multi = FALSE,
  chatty = TRUE,
  logfile = TRUE,
  ...
)
}
\arguments{
\item{where}{Path to your fastq files.}

\item{example}{Set to TRUE to run example}

\item{patternF}{Pattern of the forward/R1 reads. Default is Illumina standard.}

\item{patternR}{Pattern of the reverse/R2 reads. Default is Illumina standard}

\item{multi}{Should multithreading be done? Can use this to set to TRUE or FALSE for entire wrapper.}

\item{chatty}{How chatty should code be? Default is TRUE, but set to FALSE if you don't want so much text going to the console.}

\item{logfile}{Write a logfile to user's computer. Default is TRUE}

\item{...}{Allows passing of arguments to nested functions}
}
\value{
dada2 ASV table
}
\description{
dada2 workflow to generate ASV table
}
\examples{
dada2_asvtable(example = TRUE, logfile = FALSE)
}
