% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_decontam_wrapper.R
\name{dada2_decontam_wrapper}
\alias{dada2_decontam_wrapper}
\title{Wrapper of entire dada2/decontam process}
\usage{
dada2_decontam_wrapper(example = FALSE, passedargs = FALSE, ...)
}
\arguments{
\item{example}{Set to TRUE to run example}

\item{passedargs}{List passed arguments. Using only for troubleshooting during development; shouldn't have any use to end user}

\item{...}{Arguments to pass to nested functions}
}
\value{
A list of the new ASV table, taxa table, and metadata object
}
\description{
Wrapper of entire dada2/decontam process
}
\examples{
dada2_decontam_wrapper(example = TRUE)
}
