% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_taxa.R
\name{dada2_taxa}
\alias{dada2_taxa}
\title{dada2 taxa}
\usage{
dada2_taxa(
  asvtable = NULL,
  train = NULL,
  species = NULL,
  chatty = TRUE,
  multi = FALSE,
  example = FALSE,
  output_file = NULL,
  ...
)
}
\arguments{
\item{asvtable}{A file containing your sequences; most commonly the ASV table output from dada2.}

\item{train}{Path to database that you'd like to use for taxonomic assignment.}

\item{species}{OPTIONAL: path to database if you'd like to use assignSpecies() function}

\item{chatty}{Set to FALSE for less text to print to screen}

\item{multi}{Set to TRUE to use mutithreading where possible}

\item{example}{Run example data. Default is FALSE}

\item{output_file}{For non-example runs, path to where you'd like output files to be stored. Defaults to a folder where your fastq files are stored.}

\item{...}{Allows passing of arguments to nested functions}
}
\value{
A tibble of taxonomic assigments
}
\description{
dada2 taxa
}
\examples{
asvtable <- dada2_asvtable(example = TRUE)
train <- "inst/extdata/db/EXAMPLE_silva_nr99_v138.2_toGenus_trainset.fa.gz"
species <- "inst/extdata/db/EXAMPLE_silva_v138.2_assignSpecies.fa.gz"
dada2_taxa(asvtable = asvtable, train = train, species = species)
}
