% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_wrapper.R
\name{dada2_wrapper}
\alias{dada2_wrapper}
\title{dada2 wrapper}
\usage{
dada2_wrapper(
  example = FALSE,
  metadata = NULL,
  listargs = FALSE,
  full.wrapper = FALSE,
  log = TRUE,
  ...
)
}
\arguments{
\item{example}{Set to TRUE to run an example}

\item{metadata}{OPTIONAL. Pass associated metadata to include in output list.}

\item{listargs}{List passed arguments (dev/troubleshooting)}

\item{full.wrapper}{FALSE by default; must be FALSE when using
dada2_decontam_wrapper as otherwise metadata objects collide}

\item{log}{Produce a log file (currently not wired through)}

\item{...}{Allows passing of arguments to nested functions}
}
\value{
a list with ASV table + taxonomy table (+ metadata, optionally)
}
\description{
dada2 wrapper
}
\examples{
dada2_wrapper(example = TRUE)
}
