% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decontam_wrapper.R
\name{decontam_wrapper}
\alias{decontam_wrapper}
\title{Decontam}
\usage{
decontam_wrapper(asvtable = NULL, taxa = NULL, metadata = NULL, ...)
}
\arguments{
\item{asvtable}{asvtable}

\item{taxa}{taxa}

\item{metadata}{metadata}

\item{...}{allow arguments to be passed to nested functions}
}
\value{
asvtable with potential contaminants removed
}
\description{
Decontam
}
\examples{
asvtable <- dada2_asvtable("example")
metadata <- example_metadata()
decontam_wrapper(asvtable = asvtable, metadata = metadata)
}
