% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro4R-utils.R
\name{example_metadata}
\alias{example_metadata}
\title{Load in example metadata file}
\usage{
example_metadata()
}
\value{
An example metadata file
}
\description{
Load in example metadata file
}
\examples{
metadata <- example_metadata()
}
