% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{List of all three example objects in one package}
\format{
\subsection{\code{exampleData}}{

A list with of three tibbles:
\describe{
\item{exampleMetadata}{Example metadata tibble}
\item{exampleASVtable}{Example ASV count table}
\item{exampleTaxa}{Example ASV taxonomy table}

...
}
}
}
\source{
made up data following typical dada2 ASV count and taxonomy table style
}
\usage{
exampleData
}
\description{
Made up metadata, ASV count, and ASV taxonomy tables
}
\keyword{datasets}
