% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro4R-utils.R
\name{filtering}
\alias{filtering}
\title{Filter the ASV table}
\usage{
filtering(asvtable = NULL, minDepth = 1000, minASVCount = 2)
}
\arguments{
\item{asvtable}{Input ASV table}

\item{minDepth}{Minimum number of reads a sample must have to be kept}

\item{minASVCount}{Minimum count an ASV must have to be kept}
}
\value{
A filtered ASV table
}
\description{
Filter the ASV table
}
