% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro4R-utils.R
\name{findUserCD}
\alias{findUserCD}
\title{Collection of helper/utility functions not intended to be directly used by end user}
\usage{
findUserCD()
}
\value{
Path of user's current working directory
}
\description{
Collection of helper/utility functions not intended to be directly used by end user
}
