% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro4R-utils.R
\name{full_example_data}
\alias{full_example_data}
\title{Title}
\usage{
full_example_data(path = NULL)
}
\arguments{
\item{path}{Path to directory where you want to download file to.}
}
\value{
Downloads zipped folder of full fastq files used in example
}
\description{
Title
}
\examples{
full_example_data("example")
}
