% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro4R-utils.R
\name{known_dada2_dbs}
\alias{known_dada2_dbs}
\title{A simple tibble of the names of maintained dada2-formatted taxonomy reference databases}
\usage{
known_dada2_dbs(what = "list")
}
\arguments{
\item{what}{A keypair listing the expected pattern name or a vector of the known names.}
}
\value{
A tibble
}
\description{
A simple tibble of the names of maintained dada2-formatted taxonomy reference databases
}
\examples{
known_dada2_dbs()
}
