% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeExample.R
\name{makeExample}
\alias{makeExample}
\title{Create example data for testing and teaching purposes specifically for this package.}
\usage{
makeExample(what = NULL)
}
\arguments{
\item{what}{What do you want to create? Valid options are 'meta' (for a metadata table), 'asv' (for an ASV table), 'taxa' (for a taxonomy table) or 'all' or left blank to create a list of all three.}
}
\value{
A metadata, ASV, or taxonomy table, or a list of all three.
}
\description{
Create example data for testing and teaching purposes specifically for this package.
}
\examples{
metadata <- makeExample("meta")
asvtable <- makeExample("asv")
taxa <- makeExample("taxa")
all <- makeExample() # OR:
all <- makeExample("all")
}
