% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeExample.R
\name{makeExampleMeta}
\alias{makeExampleMeta}
\title{Create a simple metadata table}
\usage{
makeExampleMeta()
}
\value{
Simple metadata table for testing and demonstration.
}
\description{
Create a simple metadata table
}
\examples{
metadata <- makeExampleMeta()
}
