% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeExample.R
\name{makeExampleSeqtab}
\alias{makeExampleSeqtab}
\title{Creates a simple example ASV table}
\usage{
makeExampleSeqtab()
}
\value{
A simple ASV table
}
\description{
Creates a simple example ASV table
}
\examples{
makeExampleSeqtab()
}
