% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeExample.R
\name{makeExampleTaxa}
\alias{makeExampleTaxa}
\title{Create a simple taxonomy table}
\usage{
makeExampleTaxa()
}
\value{
Simple taxonomy table for testing and demonstration purposes.
}
\description{
Create a simple taxonomy table
}
\examples{
taxa <- makeExampleTaxa()
}
