% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro4R-utils.R
\name{matrixify}
\alias{matrixify}
\title{Turn a data frame or tibble into a matrix}
\usage{
matrixify(x)
}
\arguments{
\item{x}{data frame or tibble}
}
\value{
a matrix
}
\description{
Turn a data frame or tibble into a matrix
}
