% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro4R-utils.R
\name{packItUp}
\alias{packItUp}
\title{Package up your three objects (metadata, asvtable, taxonomy table) into a single list for less typing}
\usage{
packItUp(asvtable = NULL, taxa = NULL, metadata = NULL)
}
\arguments{
\item{asvtable}{ASV count table object}

\item{taxa}{Taxonomy table object}

\item{metadata}{Metadata object}
}
\value{
A list containing the above three objects
}
\description{
Package up your three objects (metadata, asvtable, taxonomy table) into a single list for less typing
}
\examples{
asvtable <- dada2_asvtable("example", logfile = FALSE)
train <- "inst/extdata/db/EXAMPLE_silva_nr99_v138.2_toGenus_trainset.fa.gz"
species <- "inst/extdata/db/EXAMPLE_silva_v138.2_assignSpecies.fa.gz"
taxa <- dada2_taxa(asvtable = asvtable, train = train, species = species)
metadata <- example_metadata()
all <- packItUp(metadata, asvtable, taxa)
}
