% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutadapt_helper.R
\name{primerHits}
\alias{primerHits}
\title{primer hits
Code was pulled from dada2 document https://benjjneb.github.io/dada2/ITS_workflow.html
Benjamin Callahan gets all the credit}
\usage{
primerHits(primer, fn)
}
\arguments{
\item{primer}{DNA sequence to search for}

\item{fn}{sequences to search in}
}
\value{
number of hits for DNA sequence of interest in fastq file
}
\description{
primer hits
Code was pulled from dada2 document https://benjjneb.github.io/dada2/ITS_workflow.html
Benjamin Callahan gets all the credit
}
