% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro4R-utils.R
\name{ref_db}
\alias{ref_db}
\title{Set a reference database for taxonomy assignment}
\usage{
ref_db(db)
}
\arguments{
\item{db}{File path to the reference database to use}
}
\value{
The file path of the database
}
\description{
Set a reference database for taxonomy assignment
}
