% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro4R-utils.R
\name{report_maker}
\alias{report_maker}
\title{Create RMarkdown report}
\usage{
report_maker(title, data, color)
}
\arguments{
\item{title}{title of the report}

\item{data}{input data}

\item{color}{color}
}
\value{
HTML report
}
\description{
Create RMarkdown report
}
