% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro4R-utils.R
\name{unsampled_example}
\alias{unsampled_example}
\title{Load in unsampled example}
\usage{
unsampled_example()
}
\value{
An list with the unsampled example metadata and asvtable files
}
\description{
Load in unsampled example
}
\examples{
unsampled_example()
}
