% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro4R-utils.R
\name{whereFastqs}
\alias{whereFastqs}
\title{Title}
\usage{
whereFastqs(path = NULL, chatty = TRUE)
}
\arguments{
\item{path}{Path to folder containing fastq files}

\item{chatty}{If set to FALSE, don't print as much to console.}
}
\value{
file path to a folder
}
\description{
Title
}
\examples{
whereFastqs(".")
}
