% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAlpha.R
\name{calcAlpha}
\alias{calcAlpha}
\title{Calculate standard alpha diversity indices}
\usage{
calcAlpha(
  asvtable = NULL,
  metadata = NULL,
  numRare = 0,
  method = "standard",
  minReads = 1
)
}
\arguments{
\item{asvtable}{ASV table}

\item{metadata}{metadata}

\item{numRare}{optional number of rarefactions to perform on ASV table}

\item{method}{doesn't do anything yet; may end up removing}

\item{minReads}{minimum number of reads to keep a sample. default is 1 to discard samples with no read, but you may set it higher if desired. if you use rarefaction, it's recommended to set this to be at least 1000.}
}
\value{
returns a tibble of the calculated values for each sample.
}
\description{
Calculate standard alpha diversity indices
}
\examples{
asvtable <- unsampled_example()$asvtable
metadata <- unsampled_example()$metadata
calcAlpha(asvtable = asvtable, metadata = metadata)
}
