% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro4R-utils.R
\name{converter}
\alias{converter}
\title{Converts tibbles, data frames, and matrices to each other as needed by the specific function}
\usage{
converter(x = NULL, out = "matrix", id = "SampleID")
}
\arguments{
\item{x}{A tibble, data frame, or matrix that you want to convert}

\item{out}{What you want to conver it to out (tibble, data_frame, or matrix)}

\item{id}{ID column, if relevant}
}
\value{
A tibble, data frame, or matrix
}
\description{
Converts tibbles, data frames, and matrices to each other as needed by the specific function
}
\examples{
converter(contaminate()$asvtable)

}
