% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_wrapper.R
\name{dada2_wrapper}
\alias{dada2_wrapper}
\title{dada2 Wrapper}
\usage{
dada2_wrapper(
  where = NULL,
  patternF = "_R1_001.fastq.gz",
  patternR = "_R2_001.fastq.gz",
  multi = FALSE,
  ...
)
}
\arguments{
\item{where}{Path to your fastq files. Run \code{dada2_wrapper(path = "example")} if you want to run the example data.}

\item{patternF}{Pattern of the forward/R1 reads. Default is Illumina standard.}

\item{patternR}{Pattern of the reverse/R2 reads. Default is Illumina standard}

\item{multi}{Should multithreading be done? Can use this to set to TRUE or FALSE for entire wrapper.}

\item{...}{Allows passing of arguments to nested functions}
}
\value{
dada2 ASV table and taxonomy table
}
\description{
dada2 Wrapper
}
\examples{
dada2_wrapper("example")
}
