% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_wrapper.R
\name{dada2_wrapper}
\alias{dada2_wrapper}
\title{dada2 wrapper}
\usage{
dada2_wrapper(metadata = NULL, ...)
}
\arguments{
\item{metadata}{OPTIONAL. Can pass your associated metadata object to this function so it packages it all up for you for downstream steps}

\item{...}{Allows passing of arguments to nested functions}
}
\value{
a list of dada2 ASV table and taxonomy table (and metadata, optionally)
}
\description{
dada2 wrapper
}
\examples{
train <- "inst/extdata/db/EXAMPLE_silva_nr99_v138.2_toGenus_trainset.fa.gz"
species <- "inst/extdata/db/EXAMPLE_silva_v138.2_assignSpecies.fa.gz"
dada2_wrapper(where = "example", train = train, species = species)
}
