% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_wrapper.R
\name{dada2_wrapper}
\alias{dada2_wrapper}
\title{Dada2 Wrapper}
\usage{
dada2_wrapper(path = ".")
}
\arguments{
\item{path}{Path to your fastq files. Run \code{dada2_wrapper(path = "example")} if you want to run the example data.}
}
\value{
dada2 ASV table and taxonomy table
}
\description{
Dada2 Wrapper
}
\examples{
dada2_wrapper()
}
