% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decontam_wrapper.R
\name{decontam_wrapper}
\alias{decontam_wrapper}
\title{Decontam}
\usage{
decontam_wrapper(asvtable = NULL, taxa = NULL, metadata = NULL, ...)
}
\arguments{
\item{asvtable}{asvtable}

\item{taxa}{taxa}

\item{metadata}{metadata}

\item{...}{allow arguments to be passed to nested functions}
}
\value{
asvtable with potential contaminants removed
}
\description{
Decontam
}
\examples{
train <- "inst/extdata/db/EXAMPLE_silva_nr99_v138.2_toGenus_trainset.fa.gz"
species <- "inst/extdata/db/EXAMPLE_silva_v138.2_assignSpecies.fa.gz"
dada2_wrapper(where = "example", train = train, species = species, logfile = FALSE)
metadata <- example_metadata()
decontam_wrapper(asvtable = asvtable, metadata = metadata, logfile = FALSE)
}
