% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decontam_wrapper.R
\name{decontam_wrapper}
\alias{decontam_wrapper}
\title{Decontam}
\usage{
decontam_wrapper(asvtable = NULL, taxa = NULL, metadata = NULL, ...)
}
\arguments{
\item{asvtable}{asvtable}

\item{taxa}{taxa}

\item{metadata}{metadata}

\item{...}{allow arguments to be passed to nested functions}
}
\value{
asvtable, taxa table, and metadata with potential contaminants removed
}
\description{
Decontam
}
\examples{
train <- "inst/extdata/db/EXAMPLE_silva_nr99_v138.2_toGenus_trainset.fa.gz"
species <- "inst/extdata/db/EXAMPLE_silva_v138.2_assignSpecies.fa.gz"

# the example ASV table is too small for decontam to work properly
# to get around this, we deliberately "contaminate" the ASV table with
# the contaminate() command. this requires us to use the matched metadata
# file created with contaminate()
asvtable <- converter(contaminate()$asvtable)
taxa <- dada2_taxa(asvtable = asvtable, train = train, species = species)
metadata <- contaminate()$metadata
decontam_wrapper(asvtable = asvtable, taxa = taxa, metadata = metadata, logfile = FALSE)
}
