% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro4R-utils.R
\name{full_example_data}
\alias{full_example_data}
\title{Option to use googledrive package to download example data}
\usage{
full_example_data(path = NULL)
}
\arguments{
\item{path}{Path to directory where you want to download file to.}
}
\value{
Downloads zipped folder of full fastq files used in example
as this is it doesn't work. may delete later.
}
\description{
Option to use googledrive package to download example data
}
