% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro4R-utils.R
\name{known_dada2_dbs}
\alias{known_dada2_dbs}
\title{Simple tibble key pair of known patterns of maintained dada2-formatted taxonomy reference databases}
\usage{
known_dada2_dbs()
}
\value{
A tibble
}
\description{
Simple tibble key pair of known patterns of maintained dada2-formatted taxonomy reference databases
}
\examples{
known_dada2_dbs()
}
