% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro4R-utils.R
\name{test_dbs}
\alias{test_dbs}
\title{Load test databases into environment without having to type so much}
\usage{
test_dbs()
}
\description{
Load test databases into environment without having to type so much
}
\examples{
db <- test_dbs()
train <- db$train
species <- db$species
}
