% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro4R-utils.R
\name{tibblefy}
\alias{tibblefy}
\title{tibblefy a specific type of data frame}
\usage{
tibblefy(x, type = "asvtable")
}
\arguments{
\item{x}{data frame you want to turn into a tibble}

\item{type}{asvtable or taxa}
}
\value{
a tibble
}
\description{
tibblefy a specific type of data frame
}
\examples{
filepath <- system.file("extdata/objects", package = "micro4R", "asvtable.csv", mustWork = TRUE)
asvtable <- read.csv(file = filepath, header=TRUE)
tibblefy(asvtable)
}
